Add "blank" react-native-babylon React Native module (#2)
This is a "blank" react-native-babylon TypeScript React Native module. Ideally it would not be checked in directly to the node_modules directory of the Playground test app, but this is the only way I could get it to work reliably. I tried hard to put it in its own directory and to get the metro bundler to work with Playground having a symbolic link under node_modules to the other directory (this is what npm install does), but it was flakey and often would fail at runtime with "HMRClient" errors. This configuration is kind of ugly, but it seems very reliable and simple, so going this route for now.

This is the second of three initial commits to create a reasonable git history from the work that has been done so far. The three commits will be:

- A "blank" Playground React Native app
- A "blank" react-native-babylon React Native module
- A fleshed out initial implementation of react-native-babylon module and Playground app to test it