Add NativeInput to the interop layer (#7)
With these changes, the interop layer in the React Native module now consumes NativeInput and forwards input events from the native SurfaceView to NativeInput. Ideally we'd find a better way to share more of this interop code through Babylon Native itself (rather than being way up here in the React Native layer), but I think we should wait and figure this out in the future.