Publish to npmjs.com (#22)
These changes update the release workflow to publish the package to npmjs instead of github. By default when publishing to npmjs packages are private, so I had to add an explicit argument to tell it to publish with public access. The NPM_TOKEN was added to the secrets of the repo by @sebavan. This should conclude the work for #6 (until/unless we decide to publish binaries instead of source).