Simplify linking for iOS (#32)
Previously, I had manually updated the Playground app to link all the necessary libs. This would have to be repeated for any other project that consumed the current source-based package. So the main change here is to move that list of libs from the consuming pbxproj to the podspec. However, to do this I needed all the lib names to be the same for debug and release, but a few bgfx related libs had a d appended to them for debug. This was fixed in the BabylonNative repo, but updating to a newer BabylonNative submodule also pulls in @syntheticmagus's shader compilation changes, which require a newer version of Babylon.js, so I also updated to the latest Babylon.js NPM package.