Explicitly update logkitty to ^0.7.1 to resolve GitHub security warning (#34)
We depend on react-native@0.62.1, which depends on @react-native-community/cli-platform-android@^4.5.1, which depends on logkitty@^0.6.0. GitHub is giving us a security warning saying we should use logkitty@^0.7.1. With this change, I'm just explicitly updating it to the newer version to override what React Native is asking for. Everything still seems to work fine after this change from some quick testing.