Remove GetJsGlobalContextRef.jar and use the checked in JSCRuntime.h to get at the JSC context (#38)
This change removes GetJsGlobalContextRef.jar (specific to Android, but derived from JSCRuntime.h) and uses the checked in JSCRuntime.h directly. It also restructures the code a bit such that the JSI runtime pointer is passed to the native layer, which matches what we do on the iOS side and better prepares for napi over jsi.