Change the Android package name to not be the default name generated by the React Native CLI (#99)
We never changed the default Android package name generated by the React Native CLI when you first create the module/package. This didn't seem to matter as it is basically an implementation detail, but it turns out if multiple packages do this, they conflict in the gradle build. So this change just gives the Android package a unique name (babylonreactnative), which should resolve this issue.